Set MySQL credentials and create Post and Comment entities
The project was created using Spring Initializr. Two classes were created in the package 'Entity', one class for the entity Post and another for the entity Comment. Then, the attributes for these to entities were defined, as well as their OneToMany relation, as any Post can relates with several comments.