Create Post and Comment Repositories
After creating a new package called repository, two repositories were added, one for the posts and another one for the comments. Next, the posts schema was created in MySQL Workbench, and after running the spring application the post and comment tables were created, as well as the table for the OneToMany relation.