Create a User entity and its relations with Post and Comment
In the new entity (User) it can be created a new user with unique email, also to get all users, to get a user by id or by email, to update a user validating if the email is available, and to delete a user by id. Now at post/comment creation, it is necessary to specify the user that creates the post/comment. When a user is deleted, all the posts and comments made by this user are deleted too. Some endpoints to get and delete posts/comments by user id were created